Try to fix migration for questionnaires # More fixes for int->long conversion

+++ src/UI/Headquarters/WB.UI.Headquarters/API/PublicApi/Models/SurveyStatisticsInput.cs
@@ -15,7 +15,7 @@ namespace WB.UI.Headquarters.API.PublicApi.Models
         public int? Max { get; set; }
         public ExportFileType? exportType { get; set; }
         public string ConditionalQuestion { get; set; }
-        public int[] Condition { get; set; }
+        public long[] Condition { get; set; }
         public bool Pivot { get; set; }
 
         public ReportMode Mode { get; set; }
